# docs/swagger.yaml
definitions:
  dto.LoginRequest:
    properties:
      email:
        example: tiago@example.com
        type: string
      password:
        example: SenhaForte123!
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      refresh_token:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.SignupRequest:
    properties:
      confirm_password:
        example: SenhaForte123!
        minLength: 6
        type: string
      email:
        example: tiago@example.com
        type: string
      name:
        example: Tiago Duarte
        maxLength: 100
        minLength: 2
        type: string
      password:
        example: SenhaForte123!
        minLength: 6
        type: string
      user_type:
        example: USER
        type: string
    required:
    - confirm_password
    - email
    - name
    - password
    - user_type
    type: object
  dto.SuccessMessage:
    properties:
      message:
        type: string
    type: object
  dto.UpdateUserRequest:
    properties:
      confirm_password:
        example: newpassword123
        type: string
      email:
        example: johndoe@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: newpassword123
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        example: "2024-03-24T15:30:00Z"
        type: string
      email:
        example: tiago@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Tiago Duarte
        type: string
      user_type:
        example: USER
        type: string
    type: object
  helper.ErrorResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: This endpoint allows an existing user to login by providing their
        email and password.
      parameters:
      - description: Login credentials
        in: body
        name: loginData
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: User Login
  /signup:
    post:
      consumes:
      - application/json
      description: This endpoint allows you to create a new user with the provided
        data.
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User registered successfully
          schema:
            $ref: '#/definitions/dto.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "409":
          description: Conflit
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Create a new user
  /users:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves a list of users with pagination support.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of users per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Get users with pagination
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: This endpoint allows an authenticated user to delete their account.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/dto.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Delete user
    get:
      consumes:
      - application/json
      description: This endpoint retrieves a user by their ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: ErrUnauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: ErrNotFound
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Get user by id
    patch:
      consumes:
      - application/json
      description: This endpoint allows a user to update their profile information.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user data
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Update user
swagger: "2.0"